+++ CHANGELOG.md
@@ -1,14 +1,23 @@
 ChangeLog
 =========
 
+v1.2.0
+------
+
+- Symfony 5.4.23
+- Replaced `evert/sitemap-php` with `presta/sitemap-bundle`
+- Replaced the `desarrolla2/cache` library with Symfony Cache Component
+
 v1.1.0
 ------
+
 - Symfony 4.4.50
 - Breaking change: upgrading to Symfony Flex, changed directory structure
 - Compatibility fixes and tests for PHP 8.2
 
 v1.0.0
 ------
+
 - Change log init
 - Symfony 3.4.49
 - ApiKit 1.8.0